FIX: farmer 삭제 수정정

+++ farmer/views.py
@@ -72,6 +72,8 @@ class FarmInfoAPIView(generics.GenericAPIView):
     # 신청서에 등록된 농지는 삭제 요청 불가능 하게 만들기
     def delete(self, request, uuid):
         land_info = self.get_object(uuid)
+        if land_info.application:
+            raise NotFound("신청서에 등록된 농지는 삭제할 수 없습니다.")
         land_info.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
